バグ修正 + 安全性強化 + UI改善 + 初回プリフェッチ最適化
セキュリティ・安全性:
- MFCallback スレッド競合修正（g_mfSessionをローカルComPtrにコピー）
- PrefetchSizesForSpread ダングリング参照修正（viewableFilesコピー）
- ShellExecute系のパスサニタイズ（ダブルクォート除去）
- VBSスクリプトのパスエスケープ
- ReadFileToBytes サイズチェック追加（0バイト/512MB超拒否）
- IsValidEntryPath 末尾..チェック追加
- GetFileSize→GetFileSizeEx統一（navbar.cpp）
- IStream*→ComPtr<IStream>（navbar.cpp）
- GetTickCount→GetTickCount64統一（viewer.cpp/app.h）
- g_settingsLoaded→atomic<bool>
- _wcsdup→AllocPathStr（ExpandTreeNode、ヒープ破壊修正）
- 書庫内メディア一時ファイルをMediaStop時に自動削除
- TOCTOU一時ファイル修正（CreateFileW CREATE_NEWループ）

バグ修正:
- 本棚モード深い階層のリビール修正（ExpandTreeNode直接呼出し+ITEM:プレフィックス有無両対応）
- 本棚モード深い階層の書庫読込修正（通常パスノードのNavigateTo追加）
- 本棚モード切替時のファイルリスト上書き防止（isRevealingガード）
- 本棚モードソートボタン表示（UpdateSortButtonLabels追加）
- 本棚モード名前変更後のツリー再構築（ShowBookshelfTree追加）
- 履歴モード左右キー対応（HistoryListサブクラスにWM_KEYDOWN転送追加）
- 履歴モード→動画再生（書庫後のフォルダ強制再読込）
- フォルダ/カテゴリ移動時のビューアークリア
- 見開きキャッシュの書庫切替時クリア
- nav.cppハードコード日本語をI18nGet置換+status.loadingキー追加
- 未使用変数names[]/keys[]削除

UI改善:
- 設定/ヘルプダイアログ背景色修正（WM_CTLCOLORSTATIC）
- 初回プリフェッチ最適化（書庫オープン直後6枚→通常12枚）
- デフォルトプリフェッチ12枚

+++ src/app.h
@@ -163,7 +163,7 @@ struct ViewerState {
     AnimState anim1, anim2;  // ページ1/2のアニメ状態
     UINT_PTR animTimer = 0;
     bool isAnimating = false;
-    DWORD lastAnimTick = 0;
+    ULONGLONG lastAnimTick = 0;
 
     // レガシーアニメーション（互換用）
     std::vector<ComPtr<ID2D1Bitmap>> animFrames;
